+++ lib/http/session.rb
@@ -108,14 +108,16 @@ module HTTP
                 response: nil, encoding: nil, follow: nil, ssl: nil, ssl_context: nil,
                 proxy: nil, nodelay: nil, features: nil, retriable: nil,
                 socket_class: nil, ssl_socket_class: nil, timeout_class: nil,
-                timeout_options: nil, keep_alive_timeout: nil, base_uri: nil, persistent: nil, &block)
+                timeout_options: nil, keep_alive_timeout: nil, base_uri: nil, persistent: nil,
+                blocklist: nil, &block)
       merged = default_options.merge(
         { headers: headers, params: params, form: form, json: json, body: body,
           response: response, encoding: encoding, follow: follow, ssl: ssl,
           ssl_context: ssl_context, proxy: proxy, nodelay: nodelay, features: features,
           retriable: retriable, socket_class: socket_class, ssl_socket_class: ssl_socket_class,
           timeout_class: timeout_class, timeout_options: timeout_options,
-          keep_alive_timeout: keep_alive_timeout, base_uri: base_uri, persistent: persistent }.compact
+          keep_alive_timeout: keep_alive_timeout, base_uri: base_uri, persistent: persistent,
+          blocklist: blocklist }.compact
       )
       client = persistent? ? nil : make_client(default_options)
       res    = perform_request(client, verb, uri, merged)